+++ modules/Sources/Dependencies/SDKSynchronizer/SDKSynchronizerTest.swift
@@ -90,7 +90,7 @@ extension SDKSynchronizerClient {
         stop: { },
         isSyncing: { false },
         isInitialized: { false },
-        importAccount: { _, _, _, _, _, _ in nil },
+        importAccount: { _, _, _, _, _, _, _ in nil },
         rewind: { _ in Empty<Void, Error>().eraseToAnyPublisher() },
         getAllTransactions: { _ in [] },
         transactionStatesFromZcashTransactions: { _, _ in [] },
@@ -146,7 +146,7 @@ extension SDKSynchronizerClient {
         stop: @escaping () -> Void = { },
         isSyncing: @escaping () -> Bool = { false },
         isInitialized: @escaping () -> Bool = { false },
-        importAccount: @escaping (String, [UInt8]?, Zip32AccountIndex?, AccountPurpose, String, String?) async throws -> AccountUUID? = { _, _, _, _, _, _ in nil },
+        importAccount: @escaping (String, [UInt8]?, Zip32AccountIndex?, AccountPurpose, String, String?, BlockHeight?) async throws -> AccountUUID? = { _, _, _, _, _, _, _ in nil },
         rewind: @escaping (RewindPolicy) -> AnyPublisher<Void, Error> = { _ in return Empty<Void, Error>().eraseToAnyPublisher() },
         getAllTransactions: @escaping (AccountUUID?) -> IdentifiedArrayOf<TransactionState> = { _ in
             let mockedCleared: [TransactionStateMockHelper] = [